댓글클래스 정의, 댓글라우터 정리
